ServiceA/requirements.txt
Django==3.2
datadog==0.40.1
ddtrace==0.48.0
django-ninja==0.12.2
JSON-log-formatter==0.3.0